Need to have name for hidden columns for snapshot restore to work
serialize/deserialize to DR has no knowledge of hidden columns, normal serialize/deserialize now has; native & stream snapshot can now save/restore hidden columns, index snapshot needs investigation

Make restore procedure DR-aware, add related tests

Consider hidden columns when loading table in EE

Move setDRTimestamp outside insertTupleCommon for snapshot restore purpose

Fix missing declaration

Refactor add hidden column logic in table converter

Add logic for EE to set uninitialized DR timestamp column during snapshot restore

Clean up unnecessary code

ENG-8723, address remaining review comments, also rebase with master branch.


ENG-8723, if schema contains hidden DR timestamp, tempTuple creation shoule initialize the hidden column value to NULL.
Update hidden DR timestamp is always allowed but insert new value to DR timestamp column need to check if it is NULL.

+++ src/ee/common/RecoveryProtoMessageBuilder.h
@@ -86,7 +86,7 @@ private:
      */
     int32_t m_tupleCount;
 
-    int32_t m_maxSerializedSize;
+    size_t m_maxSerializedSize;
 };
 }
 #endif //RECOVERY_PROTO_MESSAGE_BUILDER_
